fix executionResultsVerifier_test and improve coverage on process/common.go
